+++ CHANGELOG.md
@@ -5,6 +5,11 @@
 - Tween object implemented.
 - /between command registered to display interpolaton debug info.
 
-# (recent commits)
+# 1.1
 
 - Made the interpolation hud use less space on screen.
+- split code into multiple files inside the src/ folder, init.lua is now inteded to be used just to start execution.
+- changed the argouments of BeTween.tween(), they are documented in the docs.
+- /between command now require debug privileges and can display multiple things.
+- changed the logo.
+